# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NeoEduCore API",
        "version": "1.0.0",
        "description": "API REST del backend NeoEduCore.\n\n**Este documento se genera con `php artisan openapi:generate`**, leyendo las rutas reales de la aplicación, así que no se desincroniza al añadir o quitar endpoints. No se edita a mano.\n\nAutenticación: token opaco de Laravel Sanctum (no JWT). Multi-tenant: cada petición queda acotada a la institución del usuario autenticado.\n\nLos cuerpos de ejemplo salen de `App\\Support\\ApiSpec`, compartidos con la colección de Postman."
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Servidor"
        }
    ],
...
    "paths": {
...
        "/api/auth/login": {
            "post": {
                "tags": [
                    "01 · Auth y Sesión"
                ],
                "summary": "Login (AuthController)",
                "description": "Controlador: `App\\Http\\Controllers\\AuthController@login`. Limitado por `throttle:login`.",
                "operationId": "post_api_auth_login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            },
                            "example": {
                                "email": "admin@neoeducore.edu.co",
                                "password": "password123"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operación correcta"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "01 · Auth y Sesión"
                ],
                "summary": "Logout (AuthController)",
                "description": "Controlador: `App\\Http\\Controllers\\AuthController@logout`.",
                "operationId": "post_api_auth_logout",
                "responses": {
                    "201": {
                        "description": "Operación correcta"
                    },
                    "401": {
                        "description": "No autenticado: falta el token o ha caducado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/auth/me": {
            "get": {
                "tags": [
                    "01 · Auth y Sesión"
                ],
                "summary": "Me (AuthController)",
                "description": "Controlador: `App\\Http\\Controllers\\AuthController@me`.",
                "operationId": "get_api_auth_me",
                "responses": {
                    "200": {
                        "description": "Operación correcta"
                    },
                    "401": {
                        "description": "No autenticado: falta el token o ha caducado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/register": {
            "post": {
                "tags": [
                    "01 · Auth y Sesión"
                ],
                "summary": "Register (AuthController)",
                "description": "Controlador: `App\\Http\\Controllers\\AuthController@register`. **Roles permitidos:** admin.",
                "operationId": "post_api_register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            },
                            "example": {
                                "full_name": "Estudiante de Prueba",
                                "email": "nuevo.estudiante@neoeducore.edu.co",
                                "password": "Password123",
                                "password_confirmation": "Password123",
                                "user_type": "student"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Operación correcta"
                    },
                    "401": {
                        "description": "No autenticado: falta el token o ha caducado"
                    },
                    "403": {
                        "description": "El rol del usuario no tiene permiso, o el recurso es de otro docente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "description": "Token opaco de Laravel Sanctum. Se obtiene en `POST /api/auth/login` y se envía como `Authorization: Bearer <token>`."
            }
        }
    }
}